Allow rebar3 compilation on Windows
Since we can't assume (for rebar3 at least) that Make or gmake is
available, we instead offload the work to an escript.

The escript is a copy of the existing certs_spec/gen_certifi_mod.escript
but runs relative to the library's own root instead in order to yield
predictable results, the expected context for a rebar3 hook.